requirements.txt
Flask>=0.12.3
tweepy
python-dotenv
mysqlclient
